pyproject.toml
[project]
name = "github-db-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.1",
    "psycopg2>=2.9.10",
    "python-lorem>=1.3.0.post3",
    "sqlalchemy>=2.0.43",
    "sqlalchemy-utils>=0.42.0",
    "tqdm>=4.67.1",
]
